[Bug #7910] - MPR and Fusion do not account for varied Modality LUTs in source images [Bug #9266] - Unable to render images with very small rescale slope (<<1)  * There do not seem to be any use cases for polymorphic LUTs that can be combined in an arbitrary manner for the purposes of the image display pipeline. The standard grayscale display pipeline consists of, in sequence, a modality LUT, a renormalization function, a VOI LUT, and a color map, with well defined interfaces between successive LUTs that need not be the same. In generic ILut interface allowed these functions to be combined in a different order although, in practice, a number of restrictions built into the implementations (which already fall into separate families of modality LUTs, VOI LUTs and color maps) prevented them from being combined in arbitrary manners (e.g. color maps throw an exception if a LUT is installed after it). As a result, a number of refactorings are being undertaken to restructure the types used in the display pipeline.   * IComposableLut now provides floating-point input and output ranges.    * Base implementations IModalityLut and IVoiLut specialize these properties as integers where appropriate.

+++ ImageViewer/Imaging/ComposedLut.cs
@@ -34,8 +34,8 @@ namespace ClearCanvas.ImageViewer.Imaging
 			IComposableLut firstLut, lastLut;
 			GetFirstAndLastLut(luts, out firstLut, out lastLut, out lutCount);
 
-			_minInputValue = firstLut.MinInputValue;
-			_maxInputValue = firstLut.MaxInputValue;
+			_minInputValue = (int) firstLut.MinInputValue;
+			_maxInputValue = (int) firstLut.MaxInputValue;
 			_length = _maxInputValue - _minInputValue + 1;
 			_data = cache != null ? cache.Allocate(_length) : MemoryManager.Allocate<int>(_length);
